refactor : getSplittedWorkPlaceString->getBrandByBranchString 네이밍
근무지 포맷함수인데, 어떤 형태로 포맷이 되는데 더 직관적으로 표현하고 싶었음.
