+++ src/app.features/mypage/screens/MyPageScreen.tsx
@@ -4,7 +4,7 @@ import InfoBox from 'src/app.components/InfoBox';
 import SettingIcon from 'src/app.modules/assets/mypage/setting.svg';
 import StoreIcon from 'src/app.modules/assets/mypage/store.svg';
 import LeftArrowIcon from 'src/app.modules/assets/mypage/arrowLeft.svg';
-import { getSplittedWorkPlaceString } from 'src/app.modules/util/getSplittedWorkPlaceString';
+import { getSplittedWorkPlaceString } from 'src/app.modules/util/getBrandByBranchString ';
 import Profile from 'src/app.components/Profile';
 import { useRouter } from 'next/router';
 import { SERVICE_URL } from 'src/app.modules/constants/ServiceUrl';
